requirements.txt
discord.py
TikTokApi
playwright
